[dev.garbage] runtime: calculate allocCount
span.allocCount is maintained using allocated and as
a backup alternative allocatedMasker. The later is used
simply to check the former. allocated uses shifts to
eliminated unwanted buts while allocatedMasker uses
masks.

The approach is correct though perhaps not very performant
including using the countFree routine which does a popcount
by scaning the span's markbits adding up the total number
of free objects.

+++ src/runtime/mgcsweep.go
@@ -186,7 +186,7 @@ func (s *mspan) sweep(preserve bool) bool {
 	spc := s.spanclass
 	size := s.elemsize
 	res := false
-	nfree := 0
+	nfree := uint16(0)
 
 	c := _g_.m.mcache
 	freeToHeap := false
